requirements.txt
amqp==5.1.1
argon2==0.1.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==2.1.1
crispy-tailwind==0.5.0
cron-descriptor==1.3.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-browser-reload==1.8.0
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-environ==0.10.0
django-js-asset==2.0.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-storages==1.13.2
django-taggit==4.0.0
django-tailwind==3.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kombu==5.2.4
MarkupSafe==2.1.2
mutagen==1.46.0
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
